% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{API_GET}
\alias{API_GET}
\title{Handling API Calls}
\usage{
API_GET(
  URL,
  config = NULL,
  query = NULL,
  expected_class = NULL,
  verbose = FALSE
)
}
\arguments{
\item{URL}{the URL to be called.}

\item{config}{\code{NULL} or \code{list}, forwarded to \code{httr::GET}.}

\item{query}{\code{NULL} or \code{list}, forwarded to \code{httr::GET}.}

\item{expected_class}{\code{NULL} or character vector. If set, it is checked
if the returned object is of this class. If not, a warning will be thrown
(no error).}

\item{verbose}{logical, shows some additional information if \code{TRUE}.}
}
\value{
Returns the object we get from \code{httr::content()} after a successful
API call. If an error is detected, an error with additional details will be displayed.
}
\description{
Small helper function to handle http requests to the API.
}
\author{
Reto Stauffer
}
