% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_index.R
\docType{data}
\name{gini_index}
\alias{gini_index}
\title{Gini index of per capita household income in Brazil}
\format{
A \code{\link[dplyr:reexports]{tibble}} with 16,695 rows and 4 columns:

\describe{
\item{year}{
The census year for which the Gini index was calculated (1991, 2000, or
2010).
\cr \cr
R class: \code{\link[base:integer]{integer}}
}

\item{municipality_code}{
The unique identification code for each Brazilian municipality.
\cr \cr
R class: \code{\link[base:integer]{integer}}
}

\item{municipality}{
The name of the Brazilian municipality.
\cr \cr
R class: \code{\link[base:character]{character}}
}

\item{gini_index}{
The Gini index value for the municipality, ranging from 0 (perfect
equality) to 1 (perfect inequality).
\cr \cr
R class: \code{\link[base:double]{double}}
}
}
}
\source{
Brazilian Institute of Geography and Statistics
(\href{https://www.ibge.gov.br/}{IBGE}).
}
\usage{
gini_index
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\href{https://en.wikipedia.org/wiki/Gini_coefficient}{Gini index} of per capita
household income by Brazil's municipalities for the years 1991, 2000 and
2010.
}
\details{
The \href{https://en.wikipedia.org/wiki/Gini_coefficient}{Gini index} is a
measure of income inequality, ranging from 0 (perfect equality)
to 1 (perfect inequality). This dataset provides Gini indices for Brazilian
municipalities across three census years (1991, 2000, and 2010).

The Brazilian Institute of Geography and Statistics
(\href{https://www.ibge.gov.br/}{IBGE}) Automatic Retrieval System
(\href{https://sidra.ibge.gov.br/}{SIDRA}) provides Gini data by municipality
only for the year 1991 (\href{https://sidra.ibge.gov.br/tabela/155}{Table 155}).
However, this dataset includes information from three census periods (1991,
2000, and 2010) as processed by the Institute for Applied Economic
Research (\href{https://www.ipea.gov.br}{IPEA}), which had access to the complete
census data.

The raw data is available on Brazil's public health system (SUS) platform
\href{https://datasus.saude.gov.br/informacoes-de-saude-tabnet/}{DATASUS/Tabnet}.
It can be accessed directly at the following link:
\url{http://tabnet.datasus.gov.br/cgi/ibge/censo/cnv/ginibr.def}
\subsection{Data munging}{

The data munging process was performed in R and documented using the
\href{https://quarto.org/}{Quarto} publishing system. The entire process is fully
reproducible and available in the following GitHub repository:
\url{https://github.com/sustentarea/gs-data-munging-1}
}

\subsection{Variable naming}{

The variable names follow the
\href{https://style.tidyverse.org/syntax.html#object-names}{tidyverse style guide}
, using lowercase letters with underscores separating words (snake_case) for
clarity and consistency.
}

\subsection{Variable classes}{

No special variable classes were used in this dataset.
}
}
\references{
Instituto Brasileiro de Geografia e Estatística, Instituto de Pesquisas
Econômicas e Aplicadas, & Ministério da Saúde. (n.d.). \emph{Índice de Gini da
renda domiciliar per capita segundo município --- Período: 1991, 2000 e 2010}
[Gini index of per capita household income by municipality---period: 1991,
2000, and 2010] [Dataset]. DATASUS - Tabnet. Retrieved November 16, 2024,
from \url{http://tabnet.datasus.gov.br/cgi/ibge/censo/cnv/ginibr.def}
}
\concept{datasets}
\keyword{datasets}
