% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{gs_baseurl}
\alias{gs_baseurl}
\title{Getting API Base URL}
\usage{
gs_baseurl(version = 1L)
}
\arguments{
\item{version}{integer, defaults to 1.}
}
\value{
String, base URL for the API.
}
\description{
Just returns the base URL for API requests
}
\author{
Reto Stauffer
}
