% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gs_datasets}
\alias{gs_datasets}
\title{Getting Available Datasets}
\usage{
gs_datasets(
  type = NULL,
  mode = NULL,
  version = 1L,
  config = list(),
  verbose = FALSE
)
}
\arguments{
\item{type}{\code{NULL} or character of length \code{1} to filter the request.}

\item{mode}{\code{NULL} or character of length \code{1} to filter the request.}

\item{version}{integer, API version (defaults to 1).}

\item{config}{empty list by default; can be a named list to be fowrarded
to the \code{httr::GET} request if needed.}

\item{verbose}{logical, if set \code{TRUE} some more output will be produced.}
}
\value{
Returns a \code{data.frame} with all available data types and
API endpoints.
}
\description{
The Geosphere (ZAMG) datahub API provides an endpoint to get all
available datasets. This function returns
}
\details{
The API provides an enpoint to get all available data sets which
can be filtered using the arguments \code{type} and/or \code{mode}.
Classical usecase:

Return all data sets where \code{mode == "historical"}:
* \code{gs_datasets(mode = "historical")}

Return all data sets where \code{type == "grid"}:
* \code{gs_datasets(type = "grid")}

Can be combined (settinb both \code{type} and \code{mode}).
}
\author{
Reto Stauffer
}
