% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{gs_metadata}
\alias{gs_metadata}
\title{Downloading Dataset Meta Data}
\usage{
gs_metadata(mode, resource_id, version = 1L)
}
\arguments{
\item{mode}{character, specify mode of data.}

\item{resource_id}{character, specify resource identifier of data.}

\item{version}{integer, API version (defaults to \code{1L}).}
}
\value{
Named list with a series of information about the dataset
including the URL called, available parameters with description and unit,
available stations with unit etc.
}
\description{
Downloading Dataset Meta Data
}
\author{
Reto Stauffer
}
