% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{gs_temporal_interval}
\alias{gs_temporal_interval}
\title{Extract/Calculate Temporal Interval}
\usage{
gs_temporal_interval(x)
}
\arguments{
\item{resource_id}{character, name of the \code{resource_id}}
}
\value{
Returns an integer vector of the same length as the input
vector \code{resource_id} with the temporal resolution of the data
set in seconds or \code{NA} in case the string could not have been decoded.
}
\description{
Helper function to extract and return the temporal interval
(in seconds) for different data sets based on the \code{resource_id}
identifier. Used to estimate the number of expected values to be
retrieved as there is per-request limit.
}
\examples{
gs_temporal_interval("test-1d-string")
ds <- gs_datasets()
gs_temporal_interval(ds$resource_id)

}
\author{
Reto Stauffer
}
