% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdata-package.R
\docType{package}
\name{gsdata-package}
\alias{gsdata}
\alias{gsdata-package}
\title{gsdata: Data from Sustentarea's Global Syndemic Project}
\description{
A data package made to facilitate the documentation and sharing of data from Sustentarea's Global Syndemic Project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sustentarea.github.io/gsdata/}
  \item \url{https://github.com/sustentarea/gsdata/}
  \item Report bugs at \url{https://github.com/sustentarea/gsdata/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

Authors:
\itemize{
  \item João Pedro Junqueira Schettino \email{schettino.joaopedro@gmail.com} (\href{https://orcid.org/0009-0009-5240-7666}{ORCID})
  \item Arthur Ramalho Magalhães \email{arthurama.magalhaes@gmail.com} (\href{https://orcid.org/0000-0003-3838-2844}{ORCID})
  \item Aline Martins de Carvalho \email{alinenutri@usp.br} (\href{https://orcid.org/0000-0002-4900-5609}{ORCID})
}

Other contributors:
\itemize{
  \item Sustentarea Research and Extension Group [funder]
  \item Conselho Nacional de Desenvolvimento Científico e Tecnológico [funder]
}

}
\keyword{internal}
