% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{show_http_status_and_terminate}
\alias{show_http_status_and_terminate}
\title{Show HTTP Error Status and Terminate}
\usage{
show_http_status_and_terminate(scode, xtra = NULL)
}
\arguments{
\item{scode}{numeric, http status code.}

\item{xtra}{\code{NULL} or named list with additional information.}
}
\value{
No return, will terminate R.
}
\description{
This function is called whenever \code{httr::GET} returns an
http status code out of the \code{200} range (success).
Shows \code{\link[http]{http_status()}} code information alongside
with additional messages returned by the API (if any).
}
\author{
Reto Stauffer
}
