% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{gs_baseurl}
\alias{gs_baseurl}
\title{Getting API Base URL}
\usage{
gs_baseurl()
}
\value{
String, base URL for the API.
}
\description{
Just returns the base URL for API requests. Package developed
for API version 1; can technically be overruled by setting
\code{options("gsdata.apiversion" = 2L)}.
}
\author{
Reto Stauffer
}
