% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gs_datasets}
\alias{gs_datasets}
\title{Getting Available Datasets}
\usage{
gs_datasets(version = 1L, config = list())
}
\arguments{
\item{version}{integer, API version (defaults to 1).}

\item{config}{empty list, forwarded to \code{\link[httr]{GET}} if needed.}
}
\value{
Returns a \code{data.frame} with all available data types and
API endpoints.
}
\description{
The Geosphere (ZAMG) datahub API provides an endpoint to get all
available datasets. This function returns
}
\author{
Reto Stauffer
}
