% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{gs_datasets}
\alias{gs_datasets}
\title{Getting Available Datasets}
\usage{
gs_datasets(
  type = "station",
  mode = NULL,
  version = 1L,
  config = list(),
  verbose = FALSE
)
}
\arguments{
\item{type}{\code{NULL} or character of length \code{1} to filter the request.}

\item{mode}{\code{NULL} or character of length \code{1} to filter the request.
Currently defaults to \code{mode = "station"}.}

\item{version}{integer, API version (defaults to 1).}

\item{config}{empty list by default; can be a named list to be fowrarded
to the \code{httr::GET} request if needed.}

\item{verbose}{logical, if set \code{TRUE} some more output will be produced.}
}
\value{
Returns a \code{data.frame} with all available data types and
API endpoints. When both \code{type} and \code{mode} are equal to \code{NULL}
all data available via the API will be returned.
The most important information of this return is the \code{type}, \code{mode},
as well as \code{resource_id} which is used to perform data requests.
}
\description{
The Geosphere (ZAMG) datahub API provides an endpoint to get available
datasets. This function returns a (possibly pre-filtered) \code{data.frame}
containing the dataset \code{type}, \code{mode} and \code{resource_id} needed
to perform the data requests (see e.g., \code{gs_stationdata()}) amongst some
additional information.
}
\details{
The API provides an enpoint to get all available data sets which
can be filtered using the arguments \code{type} and/or \code{mode}.
Classical usecase:

Return all data sets where \code{mode == "historical"}:
\itemize{
\item \code{gs_datasets(mode = "historical")}
}

Return all data sets where \code{type == "grid"}:
\itemize{
\item \code{gs_datasets(type = "grid")}
}

Can be combined (settinb both \code{type} and \code{mode}).
}
\seealso{
gs_stationdata
}
\author{
Reto Stauffer
}
