% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{gs_stationdata}
\alias{gs_stationdata}
\title{Downloading Station Data}
\usage{
gs_stationdata(
  mode,
  resource_id,
  parameters,
  start,
  end,
  station_ids,
  expert = FALSE,
  version = 1L,
  format = NULL,
  verbose = FALSE
)
}
\arguments{
\item{mode}{character, specify mode of data.}

\item{resource_id}{character, specify resource identifier of data.}

\item{parameters}{character vector to define which parameters to process.}

\item{start, end}{object of class \code{Date}, \code{POSIXt}, or \code{character}.
In case of character in a non-ISO format \code{format} can be used (see below).}

\item{station_ids}{integer vector with the station IDs to be processed.}

\item{expert}{logical, defaults to \code{FALSE}. If \code{TRUE} the script will not
check if the input arguments are valid. May result in unsuccessful requests
but increases the speed as \code{gs_datasets()} and \code{gs_metadata()}
do not have to be evaluated.}

\item{version}{integer, API version (defaults to \code{1L}).}

\item{format}{\code{NULL} (default) or character string, used if \code{start}/\code{end}
are characters in a specific (non ISO) format.}

\item{verbose}{logical, if set \code{TRUE} some more output will be produced.}
}
\value{
If only data for one single station (\code{length(station_ids) == 1}) is requested,
a \code{zoo} object will be returned if data is available. If no data is available,
\code{NULL} will be returned.

When multiple stations are requested a list of \code{zoo} object (or \code{NULL} if no data
is available) is returned. The name of the list corresponds to the station id requested.
}
\description{
Accessing the API endpoint \code{v<version>/station},
see \code{\url{https://dataset.api.hub.zamg.ac.at/v1/docs/quickstart.html}}.
}
\details{
This function is a convenience function for downloading different sets of
station data from the ZAMG data hub (now Geosphere). The API may change and additional
resources may be added, for details see \code{\url{https://dataset.api.hub.zamg.ac.at/v1/docs/daten.html#available-datasets}}.

To see what's available call \code{subset(gs_datasets(), type == "station")}.
}
\author{
Reto Stauffer
}
